requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
requests
gunicorn
tensorflow-cpu